# swagger-api-v3.json
{
  "swagger": "2.0",
  "info": {
    "description": "本文档说明了每个接口返回、接收的字段信息。并且制定相应的规范作为前后端交流的渠道。",
    "version": "2.0",
    "title": "易授微服务RESTful APIs说明文档",
    "contact": {
      "name": "刘祥德",
      "url": "/",
      "email": "luvletterU@njust.edu.cn"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/arrangements/{arrangementId}/attendances": {
      "post": {
        "tags": [
          "teacher-arrangement-controller"
        ],
        "summary": "获取该门课的所有出勤信息",
        "operationId": "createAttendanceUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "attendance",
            "description": "attendance",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Attendance"
            }
          },
          {
            "name": "arrangementId",
            "in": "path",
            "description": "教师安排表的编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Attendance"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/attendances": {
      "post": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "createAttendance",
        "operationId": "createAttendanceUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "attendance",
            "description": "attendance",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Attendance"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Attendance"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "更新签到信息(可用于老师主动关闭签到)",
        "operationId": "updateAttendanceUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "attendance",
            "description": "attendance",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Attendance"
            }
          },
          {
            "name": "attendanceId",
            "in": "path",
            "description": "签到编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Attendance"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/attendances/{attendanceId}": {
      "delete": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "deleteAttendance",
        "operationId": "deleteAttendanceUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "attendanceId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Attendance"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/attendances/{attendanceId}/QR_code": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "获取签到的二维码(尚未实现)",
        "operationId": "getQRcodeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "签到编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/QRcode"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "删除签到的二维码(尚未实现)",
        "operationId": "deleteQRcodeUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "签到编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/QRcode"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/attendances/{attendanceId}/sign_records": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "获取该签到项下的所有签到记录",
        "operationId": "getAttendanceUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "签到编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SignRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "增加一条签到记录",
        "operationId": "postSignRecordUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signRecord",
            "description": "signRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          },
          {
            "name": "attendanceId",
            "in": "path",
            "description": "签到编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "scheduleId",
            "in": "query",
            "description": "课表编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/attendances/{attendanceId}/votes": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "获取该签到项下的所有投票",
        "operationId": "getVotesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "签到编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Vote"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "attendance-controller"
        ],
        "summary": "发起一个属于该签到项下的投票",
        "operationId": "createVoteUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "签到编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "vote",
            "description": "投票的模型数据",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/schedules/{scheduleId}/sign_records": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findSignRecords",
        "operationId": "findSignRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SignRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/sign_records": {
      "get": {
        "tags": [
          "sign-record-controller"
        ],
        "summary": "findAllSignRecords",
        "operationId": "findAllSignRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SignRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "sign-record-controller"
        ],
        "summary": "createSignRecord",
        "operationId": "createSignRecordUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signRecord",
            "description": "signRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "sign-record-controller"
        ],
        "summary": "updateSignRecord",
        "operationId": "updateSignRecordUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signRecord",
            "description": "signRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/votes": {
      "post": {
        "tags": [
          "vote-controller"
        ],
        "summary": "createVote",
        "operationId": "createVoteUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "vote",
            "description": "vote",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "vote-controller"
        ],
        "summary": "updateVote",
        "operationId": "updateVoteUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "vote",
            "description": "vote",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Attendance": {
      "type": "object",
      "properties": {
        "arrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        },
        "attend": {
          "type": "boolean"
        },
        "attendanceId": {
          "type": "integer",
          "format": "int32"
        },
        "closed": {
          "type": "boolean"
        },
        "closedTime": {
          "type": "string",
          "format": "date-time"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "establishedTime": {
          "type": "string",
          "format": "date-time"
        },
        "qRcode": {
          "$ref": "#/definitions/QRcode"
        }
      }
    },
    "ClassSchedule": {
      "type": "object",
      "properties": {
        "courseComment": {
          "$ref": "#/definitions/CourseComment"
        },
        "scheduleId": {
          "type": "integer",
          "format": "int32"
        },
        "student": {
          "$ref": "#/definitions/Student"
        },
        "teacherArrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        }
      }
    },
...
    "Course": {
      "type": "object",
      "properties": {
        "courseId": {
          "type": "integer",
          "format": "int32"
        },
        "courseName": {
          "type": "string"
        },
        "credit": {
          "type": "number",
          "format": "double"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "CourseComment": {
      "type": "object",
      "properties": {
        "arrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        },
        "commentId": {
          "type": "integer",
          "format": "int32"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "score": {
          "type": "number",
          "format": "float"
        }
      }
    },
...
    "Department": {
      "type": "object",
      "properties": {
        "depId": {
          "type": "integer",
          "format": "int32"
        },
        "depName": {
          "type": "string"
        }
      }
    },
...
    "QRcode": {
      "type": "object",
      "properties": {
        "codeId": {
          "type": "integer",
          "format": "int32"
        },
        "fileName": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        }
      }
    },
...
    "SignRecord": {
      "type": "object",
      "properties": {
        "attendance": {
          "$ref": "#/definitions/Attendance"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "signId": {
          "type": "integer",
          "format": "int32"
        },
        "signTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Student": {
      "type": "object",
      "properties": {
        "department": {
          "$ref": "#/definitions/Department"
        },
        "grade": {
          "type": "string"
        },
        "majorIn": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "studentGender": {
          "type": "string"
        },
        "studentId": {
          "type": "string"
        },
        "studentName": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "Teacher": {
      "type": "object",
      "properties": {
        "department": {
          "$ref": "#/definitions/Department"
        },
        "password": {
          "type": "string"
        },
        "teacherGender": {
          "type": "string"
        },
        "teacherId": {
          "type": "string"
        },
        "teacherName": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        }
      },
      "description": "教师实体"
    },
    "TeacherArrangement": {
      "type": "object",
      "properties": {
        "arrangementId": {
          "type": "integer",
          "format": "int32"
        },
        "course": {
          "$ref": "#/definitions/Course"
        },
        "place": {
          "type": "string"
        },
        "teacher": {
          "$ref": "#/definitions/Teacher"
        }
      }
    },
...
    "Vote": {
      "type": "object",
      "properties": {
        "closed": {
          "type": "boolean"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "establishedTime": {
          "type": "string",
          "format": "date-time"
        },
        "voteContent": {
          "type": "string"
        },
        "voteId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}